Use the symlink-resolved location of Python in getpath on 3.14+ (#896)
Mostly fixes #380, and also fixes #713 because we prefer the actual
binary path to argv[0].

There are two changes here. The second one is taken from PR #903 (and
indirectly from CPython upstream). The reason for that is the first
change doesn't quite work on some aarch64 configs, but the second change
doesn't cover the case of weird argv[0].

Co-authored-by: Jonathan Helmus <jjhelmus@gmail.com>